LayoutTests/platform/chromium/inspector/console/console-log-document-proto-expected.txt

CONSOLE MESSAGE: line 6: [object Object]
Test that console.dir(document.__proto__) won't result in an exception when the message is formatted in the inspector.Bug 27169.

HTMLDocument console-log-document-proto.html:6
